# api/openapi/goals-api.yaml
openapi: 3.1.0
info:
  title: SED Goals API
  version: '1.0.0'
  description: Core employee-facing API for goals, categories, KPIs, weight validation, and assignments.
servers:
  - url: /api/v1
paths:
...
  /employees/{empId}/categories/{catId}/goals:
    post:
      operationId: createGoal
      summary: Create a goal in a category
      parameters:
        - $ref: '#/components/parameters/EmpId'
        - $ref: '#/components/parameters/CatId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Created goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '400':
          $ref: '#/components/responses/InvalidGoalRequest'
        '403':
          $ref: '#/components/responses/PhaseRestricted'
        '404':
          $ref: '#/components/responses/CategoryNotFound'
        '422':
          $ref: '#/components/responses/WeightInvalid'
        '429':
          $ref: '#/components/responses/RateLimit'
  /goals/{goalId}:
    put:
      operationId: updateGoal
      summary: Update a goal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalRequest'
      responses:
        '200':
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalResponse'
        '400':
          $ref: '#/components/responses/InvalidGoalRequest'
        '403':
          $ref: '#/components/responses/PhaseRestricted'
        '404':
          $ref: '#/components/responses/GoalNotFound'
        '409':
          $ref: '#/components/responses/ConcurrentModification'
        '422':
          $ref: '#/components/responses/WeightInvalid'
    delete:
      operationId: deleteGoal
      summary: Delete a goal
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '204':
          description: Goal deleted
        '403':
          $ref: '#/components/responses/PhaseRestricted'
        '404':
          $ref: '#/components/responses/GoalNotFound'
...
  /goals/global:
    post:
      operationId: createGlobalGoal
      summary: Create a global goal
      tags: [Goals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGlobalGoalRequest'
      responses:
        '201':
          description: Created global goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalGoalResponse'
        '400':
          $ref: '#/components/responses/InvalidGoalRequest'
        '429':
          $ref: '#/components/responses/RateLimit'
    get:
      operationId: listGlobalGoals
      summary: List global goals
      tags: [Goals]
      responses:
        '200':
          description: List of global goals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalGoalListResponse'
        '429':
          $ref: '#/components/responses/RateLimit'
  /goals/global/{goalId}:
    get:
      operationId: getGlobalGoal
      summary: Get a global goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Global goal detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalGoalResponse'
        '404':
          $ref: '#/components/responses/GoalNotFound'
    put:
      operationId: updateGlobalGoal
      summary: Update a global goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGlobalGoalRequest'
      responses:
        '200':
          description: Updated global goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalGoalResponse'
        '400':
          $ref: '#/components/responses/InvalidGoalRequest'
        '404':
          $ref: '#/components/responses/GoalNotFound'
    delete:
      operationId: deleteGlobalGoal
      summary: Delete a global goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '204':
          description: Global goal deleted
        '404':
          $ref: '#/components/responses/GoalNotFound'
  /goals/global/{goalId}/execute-rules:
    post:
      operationId: executeGlobalGoalRules
      summary: Execute mass assignment rules for a global goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Assignment result
          content:
            application/json:
              schema:
                type: object
                properties:
                  assigned:
                    type: integer
        '404':
          $ref: '#/components/responses/GoalNotFound'

  # ============================================================================
  # Shared Goals
  # ============================================================================
  /goals/shared:
    post:
      operationId: createSharedGoal
      summary: Create a shared goal
      tags: [Goals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSharedGoalRequest'
      responses:
        '201':
          description: Created shared goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedGoalResponse'
        '400':
          $ref: '#/components/responses/InvalidGoalRequest'
        '429':
          $ref: '#/components/responses/RateLimit'
    get:
      operationId: listSharedGoals
      summary: List shared goals
      description: Retorna objetivos compartidos. Solo objetivos de empleados activos (is_active=true) se listan; inactivos excluidos por sync Mobonet nocturno (filtro implícito is_active).
      tags: [Goals]
      responses:
        '200':
          description: List of shared goals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedGoalListResponse'
        '429':
          $ref: '#/components/responses/RateLimit'
  /goals/shared/{goalId}:
    get:
      operationId: getSharedGoal
      summary: Get a shared goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '200':
          description: Shared goal detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedGoalResponse'
        '404':
          $ref: '#/components/responses/GoalNotFound'
    put:
      operationId: updateSharedGoal
      summary: Update a shared goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSharedGoalRequest'
      responses:
        '200':
          description: Updated shared goal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedGoalResponse'
        '400':
          $ref: '#/components/responses/InvalidGoalRequest'
        '404':
          $ref: '#/components/responses/GoalNotFound'
    delete:
      operationId: deleteSharedGoal
      summary: Delete a shared goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      responses:
        '204':
          description: Shared goal deleted
        '404':
          $ref: '#/components/responses/GoalNotFound'
  /goals/shared/{goalId}/members:
    post:
      operationId: addSharedGoalMember
      summary: Add a member to a shared goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SharedMemberResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/GoalNotFound'
  /goals/shared/{goalId}/members/{employeeId}:
    delete:
      operationId: removeSharedGoalMember
      summary: Remove a member from a shared goal
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
        - $ref: '#/components/parameters/EmployeeId'
      responses:
        '204':
          description: Member removed
        '404':
          $ref: '#/components/responses/GoalNotFound'
  /goals/shared/{goalId}/progress/{employeeId}:
    put:
      operationId: updateSharedGoalMemberProgress
      summary: Update progress for a shared goal member
      tags: [Goals]
      parameters:
        - $ref: '#/components/parameters/GoalId'
        - $ref: '#/components/parameters/EmployeeId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberProgressRequest'
      responses:
        '200':
          description: Progress updated
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '404':
          $ref: '#/components/responses/GoalNotFound'

  # ============================================================================
  # Assignments
  # ============================================================================
...
components:
  parameters:
    EmpId:
      name: empId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CatId:
      name: catId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    GoalId:
      name: goalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateGoalRequest:
      type: object
      required: [name, unit, weight, target_value]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        unit:
          type: string
          enum: [porcentaje, moneda, numero, binario]
        weight:
          type: number
          minimum: 0.01
          maximum: 100
        target_value:
          type: number
          minimum: 0
          description: "0 permitido solo si unit=binario o direction=descendente; en otros casos debe ser >0"
        direction:
          type: string
          enum: [ascendente, descendente]
          default: ascendente
        baseline_value:
          type: number
          nullable: true
          minimum: 0
          description: Required when direction is descendente; must be > target_value
        kpi_ids:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 5
    UpdateGoalRequest:
      type: object
      required: [name, unit, weight, target_value, version]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        unit:
          type: string
          enum: [porcentaje, moneda, numero, binario]
        weight:
          type: number
          minimum: 0.01
          maximum: 100
        target_value:
          type: number
          minimum: 0
          description: "0 permitido solo si unit=binario o direction=descendente; en otros casos debe ser >0"
        direction:
          type: string
          enum: [ascendente, descendente]
          default: ascendente
        baseline_value:
          type: number
          nullable: true
          minimum: 0
          description: Required when direction is descendente; must be > target_value
        version:
          type: integer
          minimum: 1
        kpi_ids:
          type: array
          items:
            type: string
            format: uuid
          maxItems: 5
    GoalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        weight:
          type: number
        target_value:
          type: number
        current_value:
          type: number
        direction:
          type: string
          enum: [ascendente, descendente]
        baseline_value:
          type: number
          nullable: true
        progress_percent:
          type: number
        state:
          type: string
          enum: [borrador, fijada, en_seguimiento, evaluada, cerrada]
        version:
          type: integer
        kpis:
          type: array
          items:
            $ref: '#/components/schemas/KpiResponse'
        pending_proposal:
          $ref: '#/components/schemas/GoalProposalResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateGlobalGoalRequest:
      type: object
      required: [name, unit, weight, target_value]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        unit:
          type: string
          enum: [porcentaje, moneda, numero, binario]
        weight:
          type: number
          minimum: 0.01
          maximum: 100
        target_value:
          type: number
          minimum: 0
          description: "0 permitido solo si unit=binario o direction=descendente; en otros casos debe ser >0"
        direction:
          type: string
          enum: [ascendente, descendente]
          default: ascendente
        baseline_value:
          type: number
          nullable: true
          minimum: 0
        goal_kind:
          type: string
        category_id:
          type: string
          format: uuid
          nullable: true
          description: "Optional; global goals typically have no category."
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/GlobalAssignmentRequest'
        rules:
          type: array
          items:
            $ref: '#/components/schemas/GlobalRuleRequest'
    UpdateGlobalGoalRequest:
      type: object
      required: [name, unit, weight, target_value]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        unit:
          type: string
          enum: [porcentaje, moneda, numero, binario]
        weight:
          type: number
          minimum: 0.01
          maximum: 100
        target_value:
          type: number
          minimum: 0
          description: "0 permitido solo si unit=binario o direction=descendente; en otros casos debe ser >0"
        direction:
          type: string
          enum: [ascendente, descendente]
          default: ascendente
        baseline_value:
          type: number
          nullable: true
          minimum: 0
        goal_kind:
          type: string
        category_id:
          type: string
          format: uuid
          nullable: true
          description: "Optional; global goals typically have no category."
...
    GlobalGoalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        direction:
          type: string
        weight:
          type: number
        target_value:
          type: number
        current_value:
          type: number
        goal_kind:
          type: string
        state:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/GlobalAssignmentResponse'
        rules:
          type: array
          items:
            $ref: '#/components/schemas/GlobalRule'
    GlobalAssignmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        goal_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        weight:
          type: number
        target_value:
          type: number
        baseline_value:
          type: number
          nullable: true
    GlobalGoalListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GlobalGoalResponse'

    # ============================================================================
    # Shared Goal Schemas
    # ============================================================================
    CreateSharedGoalRequest:
      type: object
      required: [name, unit, weight, target_value]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        unit:
          type: string
          enum: [porcentaje, moneda, numero, binario]
        weight:
          type: number
          minimum: 0.01
          maximum: 100
        target_value:
          type: number
          minimum: 0
          description: "0 permitido solo si unit=binario o direction=descendente; en otros casos debe ser >0"
        direction:
          type: string
          enum: [ascendente, descendente]
          default: ascendente
        baseline_value:
          type: number
          nullable: true
          minimum: 0
        goal_kind:
          type: string
        category_id:
          type: string
          format: uuid
          nullable: true
          description: "Optional; shared goals typically have no category."
        group_name:
          type: string
          maxLength: 255
        group_description:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/SharedMemberRequest'
    UpdateSharedGoalRequest:
      type: object
      required: [name, unit, weight, target_value]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        unit:
          type: string
          enum: [porcentaje, moneda, numero, binario]
        weight:
          type: number
          minimum: 0.01
          maximum: 100
        target_value:
          type: number
          minimum: 0
          description: "0 permitido solo si unit=binario o direction=descendente; en otros casos debe ser >0"
        direction:
          type: string
          enum: [ascendente, descendente]
          default: ascendente
        baseline_value:
          type: number
          nullable: true
          minimum: 0
        goal_kind:
          type: string
        category_id:
          type: string
          format: uuid
          nullable: true
          description: "Optional; shared goals typically have no category."
        members:
          type: array
          items:
            $ref: '#/components/schemas/SharedMemberRequest'
          description: "Optional; when present, replaces group members (delete+recreate). Omit to leave unchanged."
...
    AddMemberRequest:
      type: object
      required: [employee_id, weight, target_value]
      properties:
        employee_id:
          type: string
          format: uuid
        weight:
          type: number
          minimum: 0.01
          maximum: 100
        target_value:
          type: number
          minimum: 0
          description: "0 permitido solo si unit=binario o direction=descendente; en otros casos debe ser >0"
        baseline_value:
          type: number
          nullable: true
          minimum: 0
    UpdateMemberProgressRequest:
      type: object
      required: [current_value]
      properties:
        current_value:
          type: number
          minimum: 0
    SharedGoalResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        unit:
          type: string
        direction:
          type: string
        weight:
          type: number
        target_value:
          type: number
        current_value:
          type: number
        goal_kind:
          type: string
        state:
          type: string
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        group:
          $ref: '#/components/schemas/SharedGroupResponse'
        members:
          type: array
          items:
            $ref: '#/components/schemas/SharedMemberResponse'
...
    SharedMemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        group_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        weight:
          type: number
        target_value:
          type: number
        baseline_value:
          type: number
          nullable: true
    SharedGoalListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SharedGoalResponse'

    # ============================================================================
    # Assignment Schemas
    # ============================================================================
...
  responses:
    PhaseRestricted:
      description: Operation not allowed in current cycle phase
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    WeightInvalid:
      description: Weight validation failed (Double 100% rule)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ConcurrentModification:
      description: Optimistic locking version mismatch
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InvalidRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    CategoryNotFound:
      description: Category not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    GoalNotFound:
      description: Goal not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InvalidGoalRequest:
      description: Invalid goal request (weight, target, unit, etc.)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...